requirements.txt
flask>=2.0.0
flask-limiter>=2.0.0
python-dotenv>=0.19.0
openai>=0.27.0
psycopg2-binary>=2.9.0
python-json-logger>=2.0.0
Flask-Cors>=4.0.1
gunicorn>=20.1.0
requests>=2.28.0
tabulate>=0.9.0
colorama>=0.4.5
PyYAML>=6.0.1
jsonschema>=4.17.3
icmplib>=3.0.0
plotly==5.18.0
pandas==2.1.4 